+++ Frontend/src/main.tsx
@@ -46,7 +46,6 @@ ReactDOM.createRoot(document.getElementById('root')!).render(
             authorizationParams={{
                 redirect_uri: window.location.origin + "/home",
             }}
-            leeway={999999} // Add this line to disable the secure origin check during development
         >
             <App />
         </Auth0Provider>
