+++ Frontend/vite.config.ts
@@ -3,15 +3,14 @@ import react from '@vitejs/plugin-react';
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  define: {
-    // "process.env": process.env,
-    // // By default, Vite doesn't include shims for NodeJS/
-    // // necessary for segment analytics lib to work
-    "global": {},
-  },
+define: {
+        // "process.env": process.env,
+        // // By default, Vite doesn't include shims for NodeJS/
+        // // necessary for segment analytics lib to work
+        "global": {},
+    },
   plugins: [react()],
   optimizeDeps: {
     exclude: ['lucide-react'],
   },
-
 });
